+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/RelationsResponseDto.java
@@ -2,6 +2,7 @@ package org.grnet.cat.dtos.registry;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.grnet.cat.dtos.registry.metric.MetricResponseDto;
 
 import java.util.List;
 
@@ -11,10 +12,6 @@ public class RelationsResponseDto {
     @Schema(description = "List of Metric-Test relations associated with the motivation.")
     @JsonProperty("metric_test")
     private List<MetricTestResponseDto> metricTest;
-
-    @Schema(description = "List of Metric-Definition relations associated with the motivation.")
-    @JsonProperty("metric_definition")
-    private List<MetricDefinitionResponseDto> metricDefinition;
 
     @Schema(description = "List of Criterion-Metric relations associated with the motivation.")
     @JsonProperty("criterion_metric")
@@ -26,14 +23,6 @@ public class RelationsResponseDto {
 
     public void setMetricTest(List<MetricTestResponseDto> metricTest) {
         this.metricTest = metricTest;
-    }
-
-    public List<MetricDefinitionResponseDto> getMetricDefinition() {
-        return metricDefinition;
-    }
-
-    public void setMetricDefinition(List<MetricDefinitionResponseDto> metricDefinition) {
-        this.metricDefinition = metricDefinition;
     }
 
     public List<CriterionMetricResponseDto> getCriterionMetric() {