+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/registry/metric/MetricResponseDto.java
@@ -1,63 +1,65 @@
 package org.grnet.cat.dtos.registry.metric;
 
-import java.sql.Timestamp;
 import java.util.List;
-
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Setter;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.grnet.cat.dtos.registry.motivation.PartialMotivationResponse;
 
 public class MetricResponseDto {
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The unique identifier for the metric",
-            example = "pid_graph:1A2B3C4D"
+            description = "The Metric Object.",
+            example = "pid_graph:97ECCC27"
     )
+    @JsonProperty("metric_id")
     public String id;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The metric identifier",
-            example = "MTR001"
+            description = "The Metric Name.",
+            example = "MTR1001"
     )
-    @JsonProperty("mtr")
+    @JsonProperty("metric_mtr")
     public String MTR;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "Label for the metric",
-            example = "Performance Metric"
+            description = "The Metric Label.",
+            example = "Versioning Supported"
     )
-    @JsonProperty("label")
+    @JsonProperty("metric_label")
     public String labelMetric;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "Description of the metric",
-            example = "This metric measures performance."
+            description = "The description of the metric.",
+            example = "Evidence is available that one or more versioning approaches can be implemented"
     )
-    @JsonProperty("description")
+    @JsonProperty("metric_description")
     public String descrMetric;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "URL for more information about the metric",
-            example = "http://example.com/metric"
+            description = "The Metric Version.",
+            example = "2"
     )
-    @JsonProperty("url")
-    public String urlMetric;
+    @JsonProperty("metric_version")
+    public String version;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The ID of the Type Algorithm associated with this metric",
-            example = "pid_graph:TAL1234"
+            description = "The Type Algorithm Id.",
+            example = "pid_graph:EBCEBED1"
     )
     @JsonProperty("type_algorithm_id")
     public String typeAlgorithmId;
@@ -65,8 +67,26 @@ public class MetricResponseDto {
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The ID of the Type Metric associated with this metric",
-            example = "pid_graph:TMT1234"
+            description = "The Type Algorithm Label.",
+            example = "Test = Metric"
+    )
+    @JsonProperty("type_algorithm_label")
+    public String typeAlgorithmLabel;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Type Algorithm Description.",
+            example = "The metric is the same as the test result without any modification"
+    )
+    @JsonProperty("type_algorithm_description")
+    public String typeAlgorithmDescription;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Metric Type Label.",
+            example = "pid_graph:EBCEBED1"
     )
     @JsonProperty("type_metric_id")
     public String typeMetricId;
@@ -74,36 +94,92 @@ public class MetricResponseDto {
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "Motivation associated with the metric",
-            example = "Motivation"
+            description = "The Metric Type Label.",
+            example = "Binary"
     )
-    @JsonProperty("motivation_id")
+    @JsonProperty("type_metric_label")
+    public String typeMetricLabel;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Metric Type Description.",
+            example = "pid_graph:EBCEBED1"
+    )
+    @JsonProperty("type_metric_description")
+    public String typeMetricDescription;
+
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Type Benchmark Id.",
+            example = "pid_graph:EBCEBED1"
+    )
+    @JsonProperty("type_benchmark_id")
+    public String typeBenchmarkId;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Type Benchmark label.",
+            example = "Binary-Binary"
+    )
+    @JsonProperty("type_benchmark_label")
+    public String typeBenchmarkLabel;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Type Benchmark description.",
+            example = "The benchmark maps a binary metric value to binary assessment outcome (Pass/ Fail). The inverse can also be done."
+    )
+    @JsonProperty("type_benchmark_description")
+    public String typeBenchmarkDescription;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Type Benchmark pattern.",
+            example = "M[0,1] → [Fail, Pass]"
+    )
+    @JsonProperty("type_benchmark_patter")
+    public String typeBenchmarkPatter;
+
+    @Schema(
+            type = SchemaType.STRING,
+            implementation = String.class,
+            description = "The Motivation Id.",
+            example = "pid_graph:3E109BBA"
+    )@JsonProperty("motivation_id")
     public String lodMTV;
 
     @Schema(
             type = SchemaType.STRING,
             implementation = String.class,
-            description = "The user who populated this metric",
-            example = "0000-0002-0255-5101"
+            description = "The Benchmark Value.",
+            example = "3"
     )
-    @JsonProperty(value = "populated_by")
-    public String populatedBy;
+    @JsonProperty("value_benchmark")
+    public String valueBenchmark;
 
-    @Schema(
-            type = SchemaType.STRING,
-            implementation = String.class,
-            description = "The last time this metric was touched",
-            example = "2024-08-22T14:34:00"
-    )
-    @JsonProperty("last_touch")
-    public Timestamp lastTouch;
-
+    @Setter
     @Schema(
             type = SchemaType.ARRAY,
-            implementation = TypeMetricResponseDto.class,
-            description = "List of related metrics",
-            example = "[...]"
+            implementation = PartialMotivationResponse.class,
+            description = "List of motivations related to this test."
     )
-    @JsonProperty("type_metrics")
-    public List<TypeMetricResponseDto> typeMetrics;
+    @JsonProperty("used_by_motivations")
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    public List<PartialMotivationResponse> motivations;
+
+    @Setter
+    @Schema(
+            type = SchemaType.ARRAY,
+            implementation = MetricResponseDto.class,
+            description = "List of versions of this metric."
+    )
+    @JsonProperty("metric_versions")
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    public List<MetricResponseDto> versions;
 }