+++ entity/src/main/resources/db/migration/tables/registry/V1.84__drop_metric_definition.sql
@@ -1,0 +1,8 @@
+-- ------------------------------------------------
+-- Version: v1.84
+--
+-- Description: Migration that drops Metric Definition table
+-- -------------------------------------------------
+
+-- Drop the MetricDefinition table
+DROP TABLE IF EXISTS p_Metric_Definition;