+++ mapper/src/main/java/org/grnet/cat/mappers/registry/metric/MetricMapper.java
@@ -1,12 +1,10 @@
 package org.grnet.cat.mappers.registry.metric;
 
 import org.apache.commons.lang3.StringUtils;
-import org.grnet.cat.dtos.registry.MetricDefinitionExtendedResponse;
 import org.grnet.cat.dtos.registry.metric.MetricRequestDto;
 import org.grnet.cat.dtos.registry.metric.MetricResponseDto;
 import org.grnet.cat.dtos.registry.metric.MetricUpdateDto;
-import org.grnet.cat.entities.registry.MetricDefinitionJunction;
-import org.grnet.cat.entities.registry.MetricTestJunction;
+import org.grnet.cat.dtos.registry.metric.MetricVersionRequestDto;
 import org.grnet.cat.entities.registry.metric.Metric;
 import org.mapstruct.*;
 import org.mapstruct.factory.Mappers;
@@ -21,27 +19,29 @@ public interface MetricMapper {
 
     MetricMapper INSTANCE = Mappers.getMapper(MetricMapper.class);
 
-    @Named("map")
-    @Mapping(target = "typeAlgorithmId", expression = "java(metric.getTypeAlgorithm().getId())")
-    @Mapping(target = "typeMetricId", expression = "java(metric.getTypeMetric().getId())")
-    @Mapping(target = "typeMetrics", ignore = true)
-    MetricResponseDto metricToDto(Metric metric);
-
-    @IterableMapping(qualifiedByName = "map")
+    @IterableMapping(qualifiedByName = "mapMetric")
     List<MetricResponseDto> metricToDtos(List<Metric> entities);
 
+    @Named("mapMetric")
+    @Mapping(target = "typeMetricId", expression = "java(metric.getTypeMetric().getId())")
+    @Mapping(target = "typeMetricLabel", expression = "java(metric.getTypeMetric().getLabelTypeMetric())")
+    @Mapping(target = "typeMetricDescription", expression = "java(metric.getTypeMetric().getDescTypeMetric())")
+    @Mapping(target = "typeAlgorithmId", expression = "java(metric.getTypeAlgorithm().getId())")
+    @Mapping(target = "typeAlgorithmLabel", expression = "java(metric.getTypeAlgorithm().getLabelAlgorithmType())")
+    @Mapping(target = "typeAlgorithmDescription", expression = "java(metric.getTypeAlgorithm().getDescAlgorithmType())")
+    @Mapping(target = "typeBenchmarkId", expression = "java(metric.getTypeBenchmark().getId())")
+    @Mapping(target = "typeBenchmarkLabel", expression = "java(metric.getTypeBenchmark().getLabelBenchmarkType())")
+    @Mapping(target = "typeBenchmarkDescription", expression = "java(metric.getTypeBenchmark().getDescBenchmarkType())")
+    @Mapping(target = "typeBenchmarkPatter", expression = "java(metric.getTypeBenchmark().getPattern())")
+    @Mapping(target = "versions", ignore = true)
+    MetricResponseDto metricToDto(Metric metric);
+
     @Named("mapWithExpression")
-    @Mapping(target = "id", ignore = true)
-    @Mapping(target = "populatedBy", ignore = true)
     @Mapping(target = "lastTouch", expression = "java(Timestamp.from(Instant.now()))")
     @Mapping(target = "typeAlgorithm", ignore = true)
     @Mapping(target = "typeMetric", ignore = true)
-    @Mapping(target = "lodMTV", ignore = true)
-    @Mapping(target = "upload", ignore = true)
-    @Mapping(target = "dataType", ignore = true)
-    @Mapping(target = "criteria", ignore = true)
-    @Mapping(target = "lodMTRV", ignore = true)
-    @Mapping(target = "version", ignore = true)
+    @Mapping(target = "typeBenchmark", ignore = true)
+    //@Mapping(target = "labelMetric", ignore = true)
     Metric metricToEntity(MetricRequestDto request);
 
     @Mapping(target = "MTR", expression = "java(StringUtils.isNotEmpty(request.MTR) ? request.MTR : metric.getMTR())")
@@ -59,32 +59,13 @@ public interface MetricMapper {
     @Mapping(target = "criteria", ignore = true)
     void updateMetricFromDto(MetricUpdateDto request, @MappingTarget Metric metric);
 
-    @Named("mapMetricAndDefinition")
-    @Mapping(target = "metricId", expression = "java(metric.getId())")
-    @Mapping(target = "metricMtr", expression = "java(metric.getMTR())")
-    @Mapping(target = "metricLabel", expression = "java(metric.getLabelMetric())")
-    @Mapping(target = "metricDescription", expression = "java(metric.getDescrMetric())")
-    @Mapping(target = "metricVersion", expression = "java(String.valueOf(metric.getVersion()))")
-    @Mapping(target = "typeMetricId", expression = "java(metric.getTypeMetric().getId())")
-    @Mapping(target = "typeMetricLabel", expression = "java(metric.getTypeMetric().getLabelTypeMetric())")
-    @Mapping(target = "typeMetricDescription", expression = "java(metric.getTypeMetric().getDescTypeMetric())")
-    @Mapping(target = "typeAlgorithmId", expression = "java(metric.getTypeAlgorithm().getId())")
-    @Mapping(target = "typeAlgorithmLabel", expression = "java(metric.getTypeAlgorithm().getLabelAlgorithmType())")
-    @Mapping(target = "typeAlgorithmDescription", expression = "java(metric.getTypeAlgorithm().getDescAlgorithmType())")
-    @Mapping(target = "typeBenchmarkId", expression = "java(junction.getTypeBenchmark().getId())")
-    @Mapping(target = "typeBenchmarkLabel", expression = "java(junction.getTypeBenchmark().getLabelBenchmarkType())")
-    @Mapping(target = "typeBenchmarkDescription", expression = "java(junction.getTypeBenchmark().getDescBenchmarkType())")
-    @Mapping(target = "typeBenchmarkPatter", expression = "java(junction.getTypeBenchmark().getPattern())")
-    @Mapping(target = "valueBenchmark", expression = "java(junction.getValueBenchmark())")
-    @Mapping(target = "motivationId", expression = "java(junction.getMotivation().getId())")
-    MetricDefinitionExtendedResponse metricAndDefinitionToDto(Metric metric, MetricDefinitionJunction junction);
 
-    @IterableMapping(qualifiedByName = "mapMetricAndDefinition")
-    default List<MetricDefinitionExtendedResponse> metricAndDefinitionToDtos(List<MetricDefinitionJunction> junctions) {
-        return junctions.stream()
-                .map(junction -> metricAndDefinitionToDto(junction.getMetric(), junction))
-                .collect(Collectors.toList());
-    }
+    @Named("versionMetric")
+    @Mapping(target = "typeAlgorithm", ignore = true)
+    @Mapping(target = "typeMetric", ignore = true)
+    @Mapping(target = "typeBenchmark", ignore = true)
+    @Mapping(target = "lastTouch", expression = "java(Timestamp.from(Instant.now()))")
+    Metric versionMetricToEntity(MetricVersionRequestDto request);
 
 
 }