+++ repository/src/main/java/org/grnet/cat/repositories/registry/CriterionRepository.java
@@ -113,7 +113,7 @@ public class CriterionRepository implements Repository<Criterion, String> {
                         "        t.TES,\n" +
                         "        t.labelTest,\n" +
                         "        t.descTest,\n" +
-                        "        md.valueBenchmark,\n" +
+                        "        m.valueBenchmark,\n" +
                         "        tb.labelBenchmarkType,\n" +
                         "        tm.labelTestMethod,\n"+
                         "        t.testQuestion,\n"+
@@ -123,8 +123,7 @@ public class CriterionRepository implements Repository<Criterion, String> {
                         "        tmt.labelTypeMetric\n" +
                         "    FROM\n" +
                         "        t_Type_Benchmark tb \n" +
-                        "        INNER JOIN p_Metric_Definition md ON tb.lodTBN = md.type_benchmark_lodTBN\n" +
-                        "        INNER JOIN p_Metric m ON md.metric_lodMTR = m.lodMTR\n" +
+                        "        INNER JOIN p_Metric m ON tb.lodTBN = m.lodTBN\n" +
                         "        INNER JOIN p_Metric_Test mt ON m.lodMTR = mt.metric_lodMTR\n" +
                         "        INNER JOIN p_Test t ON mt.test_lodTES = t.lodTES\n" +
                         "        INNER JOIN t_TestMethod tm ON t.lodTME = tm.lodTME\n" +
@@ -132,8 +131,7 @@ public class CriterionRepository implements Repository<Criterion, String> {
                         "        LEFT JOIN t_Type_Algorithm ta ON m.lodTAL = ta.lodTAL\n" +
                         "        LEFT JOIN t_Type_Metric tmt ON m.lodTMT = tmt.lodTMT\n" +
                         "    WHERE\n" +
-                        "        md.motivation_lodMTV = :motivationId\n" +
-                        "        AND mt.motivation_lodMTV = :motivationId\n" +
+                        "        mt.motivation_lodMTV = :motivationId\n" +
                         "        AND cm.motivation_lodMTV = :motivationId\n" +
                         "        AND cm.criterion_lodCRI = :criterionId\n" +
                         "    ORDER BY\n" +