+++ service/src/main/java/org/grnet/cat/services/registry/metric/MetricService.java
@@ -8,10 +8,10 @@ import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.core.UriInfo;
 import org.apache.commons.lang3.StringUtils;
 import org.grnet.cat.dtos.pagination.PageResource;
-import org.grnet.cat.dtos.registry.MetricDefinitionExtendedResponse;
 import org.grnet.cat.dtos.registry.metric.MetricRequestDto;
 import org.grnet.cat.dtos.registry.metric.MetricResponseDto;
 import org.grnet.cat.dtos.registry.metric.MetricUpdateDto;
+import org.grnet.cat.dtos.registry.metric.MetricVersionRequestDto;
 import org.grnet.cat.dtos.registry.metric.TypeCombinationDto;
 import org.grnet.cat.entities.registry.*;
 import org.grnet.cat.entities.registry.metric.Metric;
@@ -22,15 +22,13 @@ import org.grnet.cat.mappers.registry.MotivationMapper;
 import org.grnet.cat.mappers.registry.metric.MetricMapper;
 
 import org.grnet.cat.repositories.registry.CriterionMetricRepository;
-import org.grnet.cat.repositories.registry.MetricDefinitionRepository;
+import org.grnet.cat.repositories.registry.TypeBenchmarkRepository;
 import org.grnet.cat.repositories.registry.metric.MetricRepository;
 import org.grnet.cat.repositories.registry.metric.TypeAlgorithmRepository;
 import org.grnet.cat.repositories.registry.metric.TypeMetricRepository;
 import org.jboss.logging.Logger;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
@@ -45,12 +43,11 @@ public class MetricService {
 
     @Inject
     TypeMetricRepository typeMetricRepository;
+
+    @Inject
+    TypeBenchmarkRepository typeBenchmarkRepository;
     @Inject
     CriterionMetricRepository criterionMetricRepository;
-
-    @Inject
-    MetricDefinitionRepository metricDefinitionRepository;
-
     private static final Logger LOG = Logger.getLogger(MetricService.class);
 
     /**
@@ -59,12 +56,11 @@ public class MetricService {
      * @param id The unique ID of the Metric item.
      * @return The corresponding Metric DTO.
      */
-    public MetricDefinitionExtendedResponse getMetricById(String id) {
+    public MetricResponseDto getMetricById(String id) {
 
-        var junction = metricDefinitionRepository.fetchMetricDefinitionByMetricId(id);
-        var metric = junction.getMetric();
+        var metric = metricRepository.findById(id);
 
-        return metricResponseWithMotivations(metric, junction);
+        return metricResponseWithMotivations(metric);
     }
 
 
@@ -88,8 +84,12 @@ public class MetricService {
         metric.setPopulatedBy(userId);
         metric.setTypeAlgorithm(Panache.getEntityManager().getReference(TypeAlgorithm.class, metricRequestDto.typeAlgorithmId));
         metric.setTypeMetric(Panache.getEntityManager().getReference(TypeMetric.class, metricRequestDto.typeMetricId));
+        metric.setTypeBenchmark(Panache.getEntityManager().getReference(TypeBenchmark.class, metricRequestDto.typeBenchmarkId));
+        metric.setVersion(1);
 
         metricRepository.persist(metric);
+
+        metric.setLodMTRV(metric.getId());
 
         return MetricMapper.INSTANCE.metricToDto(metric);
     }
@@ -107,10 +107,6 @@ public class MetricService {
     public MetricResponseDto updateMetric(String id, String userId, MetricUpdateDto request) {
 
         var metric = metricRepository.findById(id);
-
-//        if(criterionMetricRepository.existMetricInStatus(id,Boolean.TRUE)){
-//            throw new ForbiddenException("No action permitted, metric exists in a published motivation");
-//        }
 
         if(StringUtils.isNotEmpty(request.MTR) && !metric.getMTR().equalsIgnoreCase(request.MTR)){
 
@@ -134,6 +130,12 @@ public class MetricService {
             metric.setTypeMetric(Panache.getEntityManager().getReference(TypeMetric.class, request.typeMetricId));
         }
 
+        if(!Objects.isNull(request.typeBenchmarkId)){
+
+            typeBenchmarkRepository.findById(request.typeBenchmarkId);
+            metric.setTypeMetric(Panache.getEntityManager().getReference(TypeMetric.class, request.typeMetricId));
+        }
+
         return MetricMapper.INSTANCE.metricToDto(metric);
     }
 
@@ -149,125 +151,111 @@ public class MetricService {
             throw new ForbiddenException("No action permitted, metric exists in a published motivation");
         }
 
+        var metric = metricRepository.findById(id);
+
+        if (metric != null && metric.getVersion() == 1 && metricRepository.countVersion(metric.getId()) > 1) {
+            throw new ForbiddenException("Cannot delete version 1 of the metric as more versions exist");
+        }
+
         return metricRepository.deleteById(id);
     }
 
     /**
-     * Retrieves a page of Metric items.
-     *
-     * @param page The index of the page to retrieve (starting from 0).
-     * @param size The maximum number of Metric items to include in a page.
-     * @param uriInfo The Uri Info for pagination links.
-     * @return A PageResource containing the Metric items in the requested page.
-     */
-    public PageResource<MetricResponseDto> getMetriclistAll(int page, int size, UriInfo uriInfo) {
-
-        var metricPage = metricRepository.fetchMetricByPage(page, size);
-        var metricDtos = MetricMapper.INSTANCE.metricToDtos(metricPage.list());
-
-        return new PageResource<>(metricPage, metricDtos, uriInfo);
-    }
-
-    /**
-     * Retrieves a paginated list of Metric items with their associated definitions.
+     * Retrieves a paginated list of Metric items.
      *
      * @param page The index of the page to retrieve (starting from 0).
      * @param size The maximum number of Metric items to include in a page.
      * @param uriInfo The UriInfo object containing request URI details for generating pagination links.
-     * @return A PageResource containing the MetricDefinitionExtendedResponse items in the requested page.
+     * @return A PageResource containing the MetricResponseDto items in the requested page.
      */
-    public PageResource<MetricDefinitionExtendedResponse> getMetricListAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {
+    public PageResource<MetricResponseDto> getMetricListAll(String search, String sort, String order, int page, int size, UriInfo uriInfo) {
 
-        var metricDefinitionPage = metricDefinitionRepository.fetchMetricAndDefinitionByPage(search, sort, order, page, size);
+        var metricPage = metricRepository.fetchMetricByPage(search, sort, order, page, size);
+        var metrics = metricPage.list();
 
-        var junctions = metricDefinitionPage.list();
-
-        if (junctions.isEmpty()) {
-            return new PageResource<>(metricDefinitionPage, List.of(), uriInfo);
+        if (metrics.isEmpty()) {
+            return new PageResource<>(metricPage, List.of(), uriInfo);
         }
 
-        var metricIds = junctions.stream()
-                .map(j -> j.getMetric().getId())
-                .distinct()
-                .collect(Collectors.toList());
-
-        var motivationsList = metricDefinitionRepository.getMotivationsForMetricIds(metricIds);
-
-        var motivationsMap = motivationsList.stream()
-                .collect(Collectors.groupingBy(
-                        result -> (String) result[0],
-                        Collectors.mapping(result -> (Motivation) result[1], Collectors.toList())
-                ));
-
-        var metricParentIds = junctions.stream()
-                .map(j -> j.getMetric().getLodMTRV())
-                .distinct()
-                .collect(Collectors.toList());
-
-        var latestMetrics = metricRepository.fetchLatestVersionMetrics(metricParentIds);
-
-        var dtoList = latestMetrics.stream()
+        var dtoList = metrics.stream()
                 .map(metric -> {
+                    var latestVersion = metric.getVersion();
+                    var versions = getMetricVersions(metric.getLodMTRV(), latestVersion);
 
-                    var junction = junctions.stream()
-                            .filter(j -> j.getMetric().getLodMTRV().equals(metric.getLodMTRV()))
-                            .findFirst()
-                            .orElseThrow(() -> new RuntimeException("Junction not found"));
+                    var metricResponse = metricResponseWithMotivations(metric);
 
-                    var dto = metricResponseWithMotivations(metric, junction);
+                    metricResponse.setVersions(versions);
 
-                    var versions = getMetricVersions(metric.getLodMTRV(), metric.getVersion());
-
-                    var motivations = motivationsMap.getOrDefault(junction.getMetric().getId(), List.of());
-
-                    dto.setMotivations(motivations.stream()
-                            .map(MotivationMapper.INSTANCE::mapPartialMotivation)
-                            .collect(Collectors.toList()));
-
-                    dto.setVersions(versions);  // Set the versions list here
-
-
-                    return dto;
+                    return metricResponse;
                 })
                 .collect(Collectors.toList());
 
-        // Return the paginated result with the list of DTOs
-        return new PageResource<>(metricDefinitionPage, dtoList, uriInfo);
+        return new PageResource<>(metricPage, dtoList, uriInfo);
     }
 
-    public List<MetricDefinitionExtendedResponse> getMetricVersions(String metricParent, Integer latestVersion) {
+    public List<MetricResponseDto> getMetricVersions(String metricParent, Integer latestVersion) {
 
-        var metricVersions = metricDefinitionRepository.fetchMetricAndDefinitionVersion(metricParent);
+        var metricVersions = metricRepository.fetchMetricAllVersions(metricParent);
 
-        return metricVersions.stream()
-                .filter(junction -> !junction.getMetric().getVersion().equals(latestVersion)) // Exclude the latest version
-                .map(junction -> {
-                    var dto = metricResponseWithMotivations(junction.getMetric(), junction);
-                    return dto;
-                })
+        var metricVersionsWithoutLatest = metricVersions.stream()
+                .filter(metric -> !metric.getVersion().equals(latestVersion)) // Filter out the latest version
                 .collect(Collectors.toList());
+
+        if (metricVersionsWithoutLatest.isEmpty()) {
+            return List.of();
+        }
+        var dtoList = metricVersionsWithoutLatest.stream()
+                .map(this::metricResponseWithMotivations)
+                .collect(Collectors.toList());
+
+        return dtoList;
+    }
+
+
+    @Transactional
+    public MetricResponseDto versionMetric(String id, String userId, MetricVersionRequestDto request) {
+
+        var parentId = metricRepository.findById(id).getLodMTRV();
+        var parentTestVersion = metricRepository.countVersion(parentId);
+
+        var childMetric = MetricMapper.INSTANCE.versionMetricToEntity(request);
+
+        childMetric.setMTR(metricRepository.findById(id).getMTR());
+        childMetric.setPopulatedBy(userId);
+        childMetric.setLodMTRV(metricRepository.findById(id).getLodMTRV());
+        childMetric.setPopulatedBy(userId);
+        childMetric.setTypeAlgorithm(Panache.getEntityManager().getReference(TypeAlgorithm.class, request.typeAlgorithmId));
+        childMetric.setTypeMetric(Panache.getEntityManager().getReference(TypeMetric.class, request.typeMetricId));
+        childMetric.setTypeBenchmark(Panache.getEntityManager().getReference(TypeBenchmark.class, request.typeBenchmarkId));
+
+        var newVersion = (int) parentTestVersion + 1 ;
+
+        childMetric.setVersion(newVersion);
+        metricRepository.persist(childMetric);
+
+        return MetricMapper.INSTANCE.metricToDto(childMetric);
     }
 
 
     /**
-     * This method takes a Metric entity, converts it to a MetricDefinitionResponseDto, retrieves and maps
+     * This method takes a Metric entity, converts it to a MetricResponseDto, retrieves and maps
      * any associated motivations, and then sets the motivations in the response.
      *
      * @param metric The Test entity to be converted and enhanced.
      * @return A PrincipleResponseDto with associated motivations.
      */
-    public MetricDefinitionExtendedResponse metricResponseWithMotivations(Metric metric, MetricDefinitionJunction junction) {
+    public MetricResponseDto metricResponseWithMotivations(Metric metric) {
 
-        var metricAndDefinitionToDto = MetricMapper.INSTANCE.metricAndDefinitionToDto(metric, junction);
+        var metricResponseToDto = MetricMapper.INSTANCE.metricToDto(metric);
 
-        var motivations = metricDefinitionRepository.getMotivationIdsByMetric(metric.getId());
+        var motivations = metricRepository.getMotivationIdsByMetric(metric.getId());
         var motivationResponses = motivations.stream()
                 .map(MotivationMapper.INSTANCE::mapPartialMotivation)
                 .collect(Collectors.toList());
 
-        metricAndDefinitionToDto.setMotivations(motivationResponses);
+        metricResponseToDto.setMotivations(motivationResponses);
 
-        return metricAndDefinitionToDto;
+        return metricResponseToDto;
     }
 
     public List<TypeCombinationDto> getMetricTypeStatistics() {