+++ CHANGELOG.md
@@ -35,6 +35,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#500](https://github.com/FC4E-CAT/fc4e-cat-api/pull/500) CAT-893 Versioning of Existing Assessment
 - [#501](https://github.com/FC4E-CAT/fc4e-cat-api/pull/501) CAT-895 Add "enabled" Column to TestMethod Table and Support Enable/Disable Functionality via API.
 - [#503](https://github.com/FC4E-CAT/fc4e-cat-api/pull/503) CAT-898 Validation Request - add list of AAI providers
+- [#507](https://github.com/FC4E-CAT/fc4e-cat-api/pull/507) CAT-903 Add "enabled" Column to Types Related to Metric Table and Support Enable/Disable Functionality via API
 
 
 ### Fix
@@ -42,12 +43,10 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#488](https://github.com/FC4E-CAT/fc4e-cat-api/pull/488) CAT-875: Merge Test and Test Definition Entities
 - [#499](https://github.com/FC4E-CAT/fc4e-cat-api/pull/499) CAT-894 ServiceConfigurationError during Async Zenodo Publishing
 - [#508](https://github.com/FC4E-CAT/fc4e-cat-api/pull/508) CAT-904 automated_group_test Not Persisted, Causing "Fully-Automated-Validation" Button to not appear on Edit
-<<<<<<< feature/CAT-837
+- [#492](https://github.com/FC4E-CAT/fc4e-cat-api/pull/492) CAT-882 Merge Metric with Definition
 - [#511](https://github.com/FC4E-CAT/fc4e-cat-api/pull/511) CAT-837 Tests in assessment should appear in an ascending order
-=======
 - [#512](https://github.com/FC4E-CAT/fc4e-cat-api/pull/512) CAT-909 NACO list information
 
->>>>>>> devel
 
 ## 2.0.0 - 2025-03-31
 ---